arm: rockchip-rk3568-evb: add hardware id detection
The RK3568 EVB uses a voltage divider to determine the hardware ID of
the board. At the moment, the voltage levels for seven EVB variants are
defined. This commit adds a late_initcall to the board code that reads
out the voltage and populates the hardware ID as environment variable.

Signed-off-by: Michael Riesch <michael.riesch@wolfvision.net>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20210705160457.7346-1-michael.riesch@wolfvision.net
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>