net: cpsw: fix ethernet bringup in case of a fixed-link phy
In case of a fixed-link phy the cpsw_adjust_link() function is
called directly in phy_device_attach() and configures the MACCONTROL
register. Right after this configuration cpsw_init() is called and
resets the MACCONTROL register.
Fix this by initializing the slave before connecting the phy device in
cpsw_open().

Signed-off-by: Oleg Karfich <oleg.karfich@wago.com>
Link: https://lore.barebox.org/1630414594-10601-1-git-send-email-oleg.karfich@wago.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>